// lib/types/pptxgenjs/core-interfaces.d.ts
export type Coord = number | `${number}%`;
export interface PositionProps {
    /**
     * Horizontal position
     * - inches or percentage
     * @example 10.25 // position in inches
     * @example '75%' // position as percentage of slide size
     */
    x?: Coord;
    /**
     * Vertical position
     * - inches or percentage
     * @example 10.25 // position in inches
     * @example '75%' // position as percentage of slide size
     */
    y?: Coord;
    /**
     * Height
     * - inches or percentage
     * @example 10.25 // height in inches
     * @example '75%' // height as percentage of slide size
     */
    h?: Coord;
    /**
     * Width
     * - inches or percentage
     * @example 10.25 // width in inches
     * @example '75%' // width as percentage of slide size
     */
    w?: Coord;
}
/**
 * Either `data` or `path` is required
 */
export interface DataOrPathProps {
    /**
     * URL or relative path
     *
     * @example 'https://onedrives.com/myimg.png` // retrieve image via URL
     * @example '/home/gitbrent/images/myimg.png` // retrieve image via local path
     */
    path?: string;
    /**
     * base64-encoded string
     * - Useful for avoiding potential path/server issues
     *
     * @example 'image/png;base64,iVtDafDrBF[...]=' // pre-encoded image in base-64
     */
    data?: string;
}
// ...
export type Margin = number | [number, number, number, number];
export type HAlign = 'left' | 'center' | 'right' | 'justify';
export type VAlign = 'top' | 'middle' | 'bottom';
export interface BorderProps {
    /**
     * Border type
     * @default solid
     */
    type?: 'none' | 'dash' | 'solid';
    /**
     * Border color (hex)
     * @example 'FF3399'
     * @default '666666'
     */
    color?: HexColor;
    /**
     * Border size (points)
     * @default 1
     */
    pt?: number;
}
export interface HyperlinkProps {
    _rId: number;
    /**
     * Slide number to link to
     */
    slide?: number;
    /**
     * Url to link to
     */
    url?: string;
    /**
     * Hyperlink Tooltip
     */
    tooltip?: string;
}
// ...
export interface ShapeFillProps {
    /**
     * Fill color
     * - `HexColor` or `ThemeColor`
     * @example 'FF0000' // hex color (red)
     * @example pptx.SchemeColor.text1 // Theme color (Text1)
     */
    color?: Color;
    /**
     * Transparency (percent)
     * - MS-PPT > Format Shape > Fill & Line > Fill > Transparency
     * - range: 0-100
     * @default 0
     */
    transparency?: number;
    /**
     * Fill type
     * @default 'solid'
     */
    type?: 'none' | 'solid';
    /**
     * Transparency (percent)
     * @deprecated v3.3.0 - use `transparency`
     */
    alpha?: number;
}
// ...
export interface ObjectNameProps {
    /**
     * Object name
     * - used instead of default "Object N" name
     * - PowerPoint: Home > Arrange > Selection Pane...
     * @since v3.10.0
     * @default 'Object 1'
     * @example 'Antenna Design 9'
     */
    objectName?: string;
}
export interface ThemeProps {
    /**
     * Headings font face name
     * @example 'Arial Narrow'
     * @default 'Calibri Light'
     */
    headFontFace?: string;
    /**
     * Body font face name
     * @example 'Arial'
     * @default 'Calibri'
     */
    bodyFontFace?: string;
}
export type MediaType = 'audio' | 'online' | 'video';
export interface ImageProps extends PositionProps, DataOrPathProps, ObjectNameProps {
    /**
     * Alt Text value ("How would you describe this object and its contents to someone who is blind?")
     * - PowerPoint: [right-click on an image] > "Edit Alt Text..."
     */
    altText?: string;
    /**
     * Flip horizontally?
     * @default false
     */
    flipH?: boolean;
    /**
     * Flip vertical?
     * @default false
     */
    flipV?: boolean;
    hyperlink?: HyperlinkProps;
    /**
     * Placeholder type
     * - values: 'body' | 'header' | 'footer' | 'title' | et. al.
     * @example 'body'
     * @see https://docs.microsoft.com/en-us/office/vba/api/powerpoint.ppplaceholdertype
     */
    placeholder?: string;
    /**
     * Image rotation (degrees)
     * - range: -360 to 360
     * @default 0
     * @example 180 // rotate image 180 degrees
     */
    rotate?: number;
    /**
     * Enable image rounding
     * @default false
     */
    rounding?: boolean;
    /**
     * Shadow Props
     * - MS-PPT > Format Picture > Shadow
     * @example
     * { type: 'outer', color: '000000', opacity: 0.5, blur: 20,  offset: 20, angle: 270 }
     */
    shadow?: ShadowProps;
    /**
     * Image sizing options
     */
    sizing?: {
        /**
         * Sizing type
         */
        type: 'contain' | 'cover' | 'crop';
        /**
         * Image width
         * - inches or percentage
         * @example 10.25 // position in inches
         * @example '75%' // position as percentage of slide size
         */
        w: Coord;
        /**
         * Image height
         * - inches or percentage
         * @example 10.25 // position in inches
         * @example '75%' // position as percentage of slide size
         */
        h: Coord;
        /**
         * Offset from left to crop image
         * - `crop` only
         * - inches or percentage
         * @example 10.25 // position in inches
         * @example '75%' // position as percentage of slide size
         */
        x?: Coord;
        /**
         * Offset from top to crop image
         * - `crop` only
         * - inches or percentage
         * @example 10.25 // position in inches
         * @example '75%' // position as percentage of slide size
         */
        y?: Coord;
    };
    /**
     * Transparency (percent)
     * - MS-PPT > Format Picture > Picture > Picture Transparency > Transparency
     * - range: 0-100
     * @default 0
     * @example 25 // 25% transparent
     */
    transparency?: number;
}
// ...
export interface ShapeProps extends PositionProps, ObjectNameProps {
    /**
     * Horizontal alignment
     * @default 'left'
     */
    align?: HAlign;
    /**
     * Radius (only for pptx.shapes.PIE, pptx.shapes.ARC, pptx.shapes.BLOCK_ARC)
     * - In the case of pptx.shapes.BLOCK_ARC you have to setup the arcThicknessRatio
     * - values: [0-359, 0-359]
     * @since v3.4.0
     * @default [270, 0]
     */
    angleRange?: [number, number];
    /**
     * Radius (only for pptx.shapes.BLOCK_ARC)
     * - You have to setup the angleRange values too
     * - values: 0.0-1.0
     * @since v3.4.0
     * @default 0.5
     */
    arcThicknessRatio?: number;
    /**
     * Shape fill color properties
     * @example { color:'FF0000' } // hex color (red)
     * @example { color:'0088CC', transparency:50 } // hex color, 50% transparent
     * @example { color:pptx.SchemeColor.accent1 } // Theme color Accent1
     */
    fill?: ShapeFillProps;
    /**
     * Flip shape horizontally?
     * @default false
     */
    flipH?: boolean;
    /**
     * Flip shape vertical?
     * @default false
     */
    flipV?: boolean;
    /**
     * Add hyperlink to shape
     * @example hyperlink: { url: "https://github.com/gitbrent/pptxgenjs", tooltip: "Visit Homepage" },
     */
    hyperlink?: HyperlinkProps;
    /**
     * Line options
     */
    line?: ShapeLineProps;
    /**
     * Points (only for pptx.shapes.CUSTOM_GEOMETRY)
     * - type: 'arc'
     * - `hR` Shape Arc Height Radius
     * - `wR` Shape Arc Width Radius
     * - `stAng` Shape Arc Start Angle
     * - `swAng` Shape Arc Swing Angle
     * @see http://www.datypic.com/sc/ooxml/e-a_arcTo-1.html
     * @example [{ x: 0, y: 0 }, { x: 10, y: 10 }] // draw a line between those two points
     */
    points?: Array<{
        x: Coord;
        y: Coord;
        moveTo?: boolean;
    } | {
        x: Coord;
        y: Coord;
        curve: {
            type: 'arc';
            hR: Coord;
            wR: Coord;
            stAng: number;
            swAng: number;
        };
    } | {
        x: Coord;
        y: Coord;
        curve: {
            type: 'cubic';
            x1: Coord;
            y1: Coord;
            x2: Coord;
            y2: Coord;
        };
    } | {
        x: Coord;
        y: Coord;
        curve: {
            type: 'quadratic';
            x1: Coord;
            y1: Coord;
        };
    } | {
        close: true;
    }>;
    /**
     * Rounded rectangle radius (only for pptx.shapes.ROUNDED_RECTANGLE)
     * - values: 0.0 to 1.0
     * @default 0
     */
    rectRadius?: number;
    /**
     * Rotation (degrees)
     * - range: -360 to 360
     * @default 0
     * @example 180 // rotate 180 degrees
     */
    rotate?: number;
    /**
     * Shadow options
     * TODO: need new demo.js entry for shape shadow
     */
    shadow?: ShadowProps;
    /**
     * @deprecated v3.3.0
     */
    lineSize?: number;
    /**
     * @deprecated v3.3.0
     */
    lineDash?: 'dash' | 'dashDot' | 'lgDash' | 'lgDashDot' | 'lgDashDotDot' | 'solid' | 'sysDash' | 'sysDot';
    /**
     * @deprecated v3.3.0
     */
    lineHead?: 'arrow' | 'diamond' | 'none' | 'oval' | 'stealth' | 'triangle';
    /**
     * @deprecated v3.3.0
     */
    lineTail?: 'arrow' | 'diamond' | 'none' | 'oval' | 'stealth' | 'triangle';
    /**
     * Shape name (used instead of default "Shape N" name)
     * @deprecated v3.10.0 - use `objectName`
     */
    shapeName?: string;
}
// ...
export interface TableCellProps extends TextBaseProps {
    /**
     * Auto-paging character weight
     * - adjusts how many characters are used before lines wrap
     * - range: -1.0 to 1.0
     * @see https://gitbrent.github.io/PptxGenJS/docs/api-tables.html
     * @default 0.0
     * @example 0.5 // lines are longer (increases the number of characters that can fit on a given line)
     */
    autoPageCharWeight?: number;
    /**
     * Auto-paging line weight
     * - adjusts how many lines are used before slides wrap
     * - range: -1.0 to 1.0
     * @see https://gitbrent.github.io/PptxGenJS/docs/api-tables.html
     * @default 0.0
     * @example 0.5 // tables are taller (increases the number of lines that can fit on a given slide)
     */
    autoPageLineWeight?: number;
    /**
     * Cell border
     */
    border?: BorderProps | [BorderProps, BorderProps, BorderProps, BorderProps];
    /**
     * Cell colspan
     */
    colspan?: number;
    /**
     * Fill color
     * @example { color:'FF0000' } // hex color (red)
     * @example { color:'0088CC', transparency:50 } // hex color, 50% transparent
     * @example { color:pptx.SchemeColor.accent1 } // theme color Accent1
     */
    fill?: ShapeFillProps;
    hyperlink?: HyperlinkProps;
    /**
     * Cell margin (inches)
     * @default 0
     */
    margin?: Margin;
    /**
     * Cell rowspan
     */
    rowspan?: number;
}
// ...
export type TableRow = TableCell[];
// ...
export interface SectionProps {
    _type: 'user' | 'default';
    _slides: PresSlide[];
    /**
     * Section title
     */
    title: string;
    /**
     * Section order - uses to add section at any index
     * - values: 1-n
     */
    order?: number;
}
// ...
export interface ObjectOptions extends ImageProps, PositionProps, ShapeProps, TableCellProps, TextPropsOptions {
    _placeholderIdx?: number;
    _placeholderType?: PLACEHOLDER_TYPE;
    cx?: Coord;
    cy?: Coord;
    margin?: Margin;
    colW?: number | number[];
    rowH?: number | number[];
}
// ...
export interface PresSlide extends SlideBaseProps {
    _rId: number;
    _slideLayout: SlideLayout;
    _slideId: number;
    addChart: (type: CHART_NAME | IChartMulti[], data: IOptsChartData[], options?: IChartOpts) => PresSlide;
    addImage: (options: ImageProps) => PresSlide;
    addMedia: (options: MediaProps) => PresSlide;
    addNotes: (notes: string) => PresSlide;
    addShape: (shapeName: SHAPE_NAME, options?: ShapeProps) => PresSlide;
    addTable: (tableRows: TableRow[], options?: TableProps) => PresSlide;
    addText: (text: string | TextProps[], options?: TextPropsOptions) => PresSlide;
    /**
     * Background color or image (`color` | `path` | `data`)
     * @example { color: 'FF3399' } - hex color
     * @example { color: 'FF3399', transparency:50 } - hex color with 50% transparency
     * @example { path: 'https://onedrives.com/myimg.png` } - retrieve image via URL
     * @example { path: '/home/gitbrent/images/myimg.png` } - retrieve image via local path
     * @example { data: 'image/png;base64,iVtDaDrF[...]=' } - base64 string
     * @since v3.3.0
     */
    background?: BackgroundProps;
    /**
     * Default text color (hex format)
     * @example 'FF3399'
     * @default '000000' (DEF_FONT_COLOR)
     */
    color?: HexColor;
    /**
     * Whether slide is hidden
     * @default false
     */
    hidden?: boolean;
    /**
     * Slide number options
     */
    slideNumber?: SlideNumberProps;
}